Reverted interesting J5 behaviour into something Javac 1.4 can deal with
git-svn-id: https://cdk.svn.sourceforge.net/svnroot/cdk/trunk/cdk@4439 eb4e18e3-b210-0410-a6ab-dec725e4b171

+++ src/org/openscience/cdk/test/qsar/CPSADescriptorTest.java
@@ -72,10 +72,10 @@ public class CPSADescriptorTest extends CDKTestCase {
             }
             */
 
-            assertEquals(0, ((Double)retval.get(28)).doubleValue(), 0.0001);
-            assertEquals(1, ((Double)retval.get(27)).doubleValue(), 0.0001);
-            assertEquals(0, ((Double)retval.get(26)).doubleValue(), 0.0001);
-            assertEquals(356.8849, ((Double)retval.get(25)).doubleValue(), 0.0001);
+            assertEquals(0, retval.get(28), 0.0001);
+            assertEquals(1, retval.get(27), 0.0001);
+            assertEquals(0, retval.get(26), 0.0001);
+            assertEquals(356.8849, retval.get(25), 0.0001);
 
         }
 }
